content: moved all CommittedReadManager methods to one file
This is completely mechanical to faciliate code review.
