+++ repo/content/content_index_recovery.go
@@ -200,35 +200,3 @@ func (bm *Manager) writePackFileIndexRecoveryData(buf *gather.WriteBuffer, pendi
 
 	return nil
 }
-
-func (bm *CommittedReadManager) readPackFileLocalIndex(ctx context.Context, packFile blob.ID, packFileLength int64) ([]byte, error) {
-	// TODO(jkowalski): optimize read when packFileLength is provided
-	_ = packFileLength
-
-	payload, err := bm.st.GetBlob(ctx, packFile, 0, -1)
-	if err != nil {
-		return nil, errors.Wrapf(err, "error getting blob %v", packFile)
-	}
-
-	postamble := findPostamble(payload)
-	if postamble == nil {
-		return nil, errors.Errorf("unable to find valid postamble in file %v", packFile)
-	}
-
-	if uint64(postamble.localIndexOffset+postamble.localIndexLength) > uint64(len(payload)) {
-		// invalid offset/length
-		return nil, errors.Errorf("unable to find valid local index in file %v", packFile)
-	}
-
-	encryptedLocalIndexBytes := payload[postamble.localIndexOffset : postamble.localIndexOffset+postamble.localIndexLength]
-	if encryptedLocalIndexBytes == nil {
-		return nil, errors.Errorf("unable to find valid local index in file %v", packFile)
-	}
-
-	localIndexBytes, err := bm.decryptAndVerify(encryptedLocalIndexBytes, postamble.localIndexIV)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to decrypt local index")
-	}
-
-	return localIndexBytes, nil
-}
